+++ backend/configuration/production.yaml
@@ -1,5 +1,4 @@
 application:
-  # I don't think base_url is used anywhere so just include
   base_url: "http://127.0.0.1"
   host: 0.0.0.0
 database:
