+++ backend/src/authentication/mod.rs
@@ -1,4 +1,5 @@
 //! backend/src/authentication/mod.rs
+//! backend/src/authentication/mod.rs
 //! Holds logic and helpers related to authenticating users.
 pub mod middleware;
 pub mod password;
